+++ CHANGELOG.rst
@@ -9,6 +9,10 @@ This project adheres to `Semantic Versioning <https://semver.org/>`_.
 
 **Added**
 
+* Link offers to project now. The ``life.qbic.business.offers.Offer`` and ``life.qbic.portal
+.offermanager.dataresources.offers`` have been extended with a new property to associate it with
+an existing project by its project identifier.
+
 **Fixed**
 
 **Dependencies**